Moved event_grade to canvas folder and implementing aggregate functions for the calculation
